DROOLS-5317-2: Tests & Minors
