+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/main/java/org/drools/workbench/screens/scenariosimulation/client/rightpanel/ListGroupItemPresenter.java
@@ -198,8 +198,8 @@ public class ListGroupItemPresenter implements ListGroupItemView.Presenter {
             toPopulate.setFactNameAndType(factName, factModelTree.getFactName()); // the name of the property differ from the type of the factModelTree: this means that we are populating children of the class
         }
         String fullPath = parentPath.isEmpty() ? factName : parentPath + "." + factName;
-        factModelTree.getSimpleProperties().forEach(
-                (key, value) -> toPopulate.addFactField(fieldItemPresenter.getLIElement(fullPath, factName, key, value.getTypeName(), value.getPropertyTypeNameToVisualize())));
+        factModelTree.getSimpleProperties().forEach((key, value) ->
+                toPopulate.addFactField(fieldItemPresenter.getLIElement(fullPath, factName, key, value.getTypeName(), value.getPropertyTypeNameToVisualize())));
         factModelTree.getExpandableProperties().forEach(
                 (key, value) -> toPopulate.addExpandableFactField(getDivElement(fullPath, key, value)));
     }
